type(reduction) dataframe
This was a badly conceived idea: to have a common reduced type
for both global and city localities.

The new idea is to have two different types for the reduction.
For global the type will be dict[Month] sortedDict.

For city the type will be dict[city_name]dict[Month] sortedDict.